Added Firebase Integration & Tested
