+++ app/build.gradle
@@ -31,6 +31,11 @@ android {
     }
 }
 
+apply plugin: 'com.android.application'
+
+// Add this line
+apply plugin: 'com.google.gms.google-services'
+
 dependencies {
 
     implementation 'androidx.core:core-ktx:1.7.0'
@@ -40,4 +45,17 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+	
+    // Import the Firebase BoM
+	implementation platform('com.google.firebase:firebase-bom:29.1.0')
+
+
+	// Add the dependency for the Firebase SDK for Google Analytics
+	// When using the BoM, don't specify versions in Firebase dependencies
+	implementation 'com.google.firebase:firebase-analytics-ktx'
+	implementation 'com.google.firebase:firebase-database-ktx'
+
+
+	// Add the dependencies for any other desired Firebase products
+	// https://firebase.google.com/docs/android/setup#available-libraries
 }
