+++ app/google-services.json
@@ -1,0 +1,48 @@
+{
+  "project_info": {
+    "project_number": "223236194053",
+    "firebase_url": "https://resresapp-default-rtdb.asia-southeast1.firebasedatabase.app",
+    "project_id": "resresapp",
+    "storage_bucket": "resresapp.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:223236194053:android:4a6f21323e2ea707f9816c",
+        "android_client_info": {
+          "package_name": "iit.heathens.reservationapp"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "223236194053-g3fiuufq1b8f2uud4ksg64bi8cfc5pgt.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "iit.heathens.reservationapp",
+            "certificate_hash": "8b40e1c6e040f28b7ebdf5ffb3d9b5750c726f91"
+          }
+        },
+        {
+          "client_id": "223236194053-starc7o7t6ac91t80r6ge5j4361p4lcc.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyALDaHnnmZ8xO388_Jtmcsxnf-Z-zKct9I"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "223236194053-starc7o7t6ac91t80r6ge5j4361p4lcc.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
