+++ app/src/main/java/iit/heathens/reservationapp/IntroPage.java
@@ -1,0 +1,23 @@
+package iit.heathens.reservationapp;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+public class IntroPage extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_intro);
+
+
+        // Write a message to the database
+        FirebaseDatabase database = FirebaseDatabase.getInstance();
+        DatabaseReference myRef = database.getReference("/Python AI Server/Predictions/");
+
+        myRef.setValue("Hello, World!");
+    }
+}
