--- app/src/main/java/iit/heathens/reservationapp/activity_register.java
@@ -1,14 +1,0 @@
-package iit.heathens.reservationapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class activity_register extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
-    }
-}
