+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background2"
-    tools:context=".activity_login">
+    tools:context=".LoginPage">
 
     <EditText
         android:id="@+id/loginEmail"
