+++ app/src/main/res/layout/activity_prototype_tester.xml
@@ -4,15 +4,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity_prototype_tester">
-
-    <!-- Opens from the activity_startup_page.xml where the Open Prototype Tester Button is clicked -->
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Button"
-        tools:layout_editor_absoluteX="89dp"
-        tools:layout_editor_absoluteY="226dp" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+    tools:context=".PrototypeTester"/>
