+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background2"
-    tools:context=".activity_register">
+    tools:context=".RegisterPage">
 
     <EditText
         android:id="@+id/fullName"
