+++ app/src/main/res/layout/activity_startup_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity_startup_page">
+    tools:context=".StartupPage">
 
     <!-- Opens from the acitivity_intro.xml which contained the Get Started Button -->
     <!-- This page includes three selections; Open Prototype Tester, Open Customer App, and Open Restaurant Management App. -->
