+++ build.gradle
@@ -1,4 +1,17 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
+buildscript {
+  repositories {
+    // Check that you have the following line (if not, add it):
+    gradlePluginPortal()
+	google()	// Google's Maven repository
+	mavenCentral()  
+  }
+  dependencies {
+    // Add this line
+    classpath 'com.google.gms:google-services:4.3.10'
+  }
+}
+
 plugins {
     id 'com.android.application' version '7.1.1' apply false
     id 'com.android.library' version '7.1.1' apply false
@@ -7,4 +20,13 @@ plugins {
 
 task clean(type: Delete) {
     delete rootProject.buildDir
+}
+
+allprojects {
+  repositories {
+    // Check that you have the following line (if not, add it):
+    gradlePluginPortal()
+	google()	// Google's Maven repository
+	mavenCentral()  
+  }
 }
